Move code to seperate module
